Refactor: Organize parser test into folders
