+++ tests/parser_test.py
@@ -9,85 +9,6 @@ from yalul.lex.token_type import TokenType
 class TestParserBinary:
     """Test parser generating binary operations expressions"""
 
-    def test_parser_run_generates_correct_ast_single_binary_expression_sum(self):
-        """
-        Validates if parser is generating a correct AST to a single binary operation sum
-        """
-        tokens = [
-            Token(TokenType.INTEGER, 2),
-            Token(TokenType.SUM, "+"),
-            Token(TokenType.INTEGER, 3),
-            Token(TokenType.EOF, "End of File")
-        ]
-
-        ast = Parser(tokens).parse()
-        ast = ast[0]
-
-        assert type(ast) is Binary
-        assert ast.operator.type == TokenType.SUM
-
-        assert type(ast.left) == Value
-        assert ast.left.value == 2
-
-        assert ast.right.value == 3
-        assert type(ast.right) == Value
-
-    def test_parser_run_generates_correct_ast_complex_binary_expression_sum(self):
-        """
-        Validates if parser is generating a correct AST to a complex binary operation sum, like 39 + 1 + 2
-        """
-        tokens = [
-            Token(TokenType.INTEGER, 39),
-            Token(TokenType.SUM, "+"),
-            Token(TokenType.INTEGER, 1),
-            Token(TokenType.SUM, "+"),
-            Token(TokenType.INTEGER, 2),
-            Token(TokenType.EOF, "End of File")
-        ]
-
-        ast = Parser(tokens).parse()
-
-        ast = ast[0]
-
-        assert type(ast) is Binary
-        assert ast.operator.type is TokenType.SUM
-
-        assert type(ast.left) is Value
-        assert ast.left.value == 39
-
-        assert type(ast.right) is Binary
-        assert ast.right.operator.type is TokenType.SUM
-
-        assert type(ast.right.left) is Value
-        assert ast.right.left.value == 1
-
-        assert type(ast.right.right) is Value
-        assert ast.right.right.value == 2
-
-    def test_parser_run_generates_correct_ast_single_binary_expression_multiply(self):
-        """
-        Validates if parser is generating a correct AST to a single binary operation multiply, like 2 * 1
-        """
-        tokens = [
-            Token(TokenType.INTEGER, 42),
-            Token(TokenType.MULTIPLY, "*"),
-            Token(TokenType.INTEGER, 1),
-            Token(TokenType.EOF, "End of File")
-        ]
-
-        ast = Parser(tokens).parse()
-
-        ast = ast[0]
-
-        assert type(ast) is Binary
-        assert ast.operator.type == TokenType.MULTIPLY
-
-        assert type(ast.left) == Value
-        assert ast.left.value == 42
-
-        assert ast.right.value == 1
-        assert type(ast.right) == Value
-
     def test_parser_run_generates_correct_ast_complex_binary_expression_with_multi_precedence(self):
         """
         Validates if parser is generating a correct AST to a binary expression with multi precedence, like 39 * 2 + 42
@@ -119,51 +40,3 @@ class TestParserBinary:
 
         assert type(ast.right) is Value
         assert ast.right.value is 42
-
-    def test_parser_run_generates_correct_ast_single_binary_expression_minus(self):
-        """
-        Validates if parser is generating a correct AST to a single binary operation minus
-        """
-        tokens = [
-            Token(TokenType.INTEGER, 5),
-            Token(TokenType.MINUS, "-"),
-            Token(TokenType.INTEGER, 3),
-            Token(TokenType.EOF, "End of File")
-        ]
-
-        ast = Parser(tokens).parse()
-        ast = ast[0]
-
-        assert type(ast) is Binary
-        assert ast.operator.type == TokenType.MINUS
-
-        assert type(ast.left) == Value
-        assert ast.left.value == 5
-
-        assert ast.right.value == 3
-        assert type(ast.right) == Value
-
-    def test_parser_run_generates_correct_ast_single_binary_expression_division(self):
-        """
-        Validates if parser is generating a correct AST to a single binary operation division
-        """
-        tokens = [
-            Token(TokenType.INTEGER, 10),
-            Token(TokenType.DIVISION, "/"),
-            Token(TokenType.INTEGER, 2),
-            Token(TokenType.EOF, "End of File")
-        ]
-
-        ast = Parser(tokens).parse()
-        ast = ast[0]
-
-        assert type(ast) is Binary
-        assert ast.operator.type == TokenType.DIVISION
-
-        assert type(ast.left) == Value
-        assert ast.left.value == 10
-
-        assert ast.right.value == 2
-        assert type(ast.right) == Value
-
-
